requirements.txt
Chameleon==2.11
FormEncode==1.2.6
Mako==0.7.3
MarkupSafe==0.15
## FIXME: could not find svn URL in dependency_links for this package:
Ming==0.3.2dev-20121219
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
Pygments==1.5
WebOb==1.2.3
WebTest==1.4.3
distribute==0.6.24
mock==1.0.1
nose==1.2.1
pycrypto==2.6
pymongo==2.4.1
pyramid==1.4
pyramid-debugtoolbar==1.0.4
repoze.lru==0.6
requests==1.0.4
translationstring==1.1
venusian==1.0a7
waitress==0.8.2
zope.deprecation==4.0.2
zope.interface==4.0.3
